requirements.txt
alembic==1.5.5
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pendulum==2.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytzdata==2020.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
